Rename and move session management
